moving out of cloudflare.py into dags
